+++ CHANGELOG.md
@@ -2,5 +2,10 @@
 
 ## Unreleased
 
+## Version 1.1.0 (2026-03-19)
+* Added `count`, `pluck`, `orWhere`, `whereIn`, `orWhereIn`, `whereNotIn` and `orWhereNotIn` query methods
+* Added `save` and `delete` methods for write support
+* Added model events (`creating`, `created`, `updating`, `updated`, `saving`, `saved`, `deleting` and `deleted`)
+
 ## Version 1.0.0 (2026-03-18)
 * Initial release